// Angular-Frontend/src/app/model/vehicle.ts
// import {Image} from './image';
import {Person} from './person';

export class Vehicle {
  vehicleId: number;
  // vehicleImg: Image;
  vehicleImg: string;
  vehicleListed: string;
  vehicleMake: string;
  vehicleModel: string;
  vehicleColour: string;
  licenceNo: string;
  vehicleCreated: any;
  vehicleDeleted: any;
  person: Person;
}